requirements.txt
# Core Dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# LLM and Agent Framework
langchain>=0.1.0
langchain-groq>=0.1.0
groq>=0.4.0

# Vector Database
chromadb>=0.4.22

# Data Processing and Visualization
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
pydantic>=2.6.0
tiktoken>=0.6.0

# PDF Generation
reportlab>=4.0.0

# Multi-Database Support
sqlalchemy>=2.0.0
pymysql>=1.1.0           # MySQL support
psycopg2-binary>=2.9.0   # PostgreSQL support  
# pyodbc>=5.0.0          # SQL Server support (uncomment if needed)
# cx-Oracle>=8.3.0       # Oracle support (uncomment if needed)
